Refactor Team database access and introduce TeamDao

+++ app/src/main/java/com/sedilant/yambol/UpdateTeamUseCaseImpl.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 class UpdateTeamUseCaseImpl @Inject constructor(
     private val teamRepository: TeamRepository
 ) : UpdateTeamUseCase {
-    override suspend fun invoke(teamId: Int, newName: String) {
+    override suspend fun invoke(teamId: Long, newName: String) {
         teamRepository.updateTeam(teamId, newName.lowercase())
     }
 }
